enable server-side bucket to bucket copies
The 'copy_object' method permits setting different source and target buckets. When different, a server-side copy takes place between the buckets.

Appropriate ACLs must exist in AWS, but this is a quick method for copying keys between buckets within the same account and region.

Copying across regions and/or accounts is more complex, and is not enabled with this change.