Add ContentType Detection with fallback
The boto3 `upload_fileobj` never guesses the ContentType (unlike other Boto3 methods). So all files are stored as `binary/octet-stream`.

This can problematic when uploading images or audio, and expecting it to load inline.

`guess_type` is a lightweight guesser, that only looks at filenames. If this implementation fails to identify a type, it falls back to the defaults `binary/octet-stream`